coinInfoFrag added :)

+++ app/src/main/java/com/camo/kripto/data/model/MarketChart.kt
@@ -1,7 +1,7 @@
 package com.camo.kripto.data.model
 
 data class MarketChart(
-    val prices: List<List<Long>>,
-    val market_caps: List<List<Long>>,
-    val total_volumes: List<List<Long>>
+    val prices: List<List<Double>>,
+    val market_caps: List<List<Double>>,
+    val total_volumes: List<List<Double>>
 )
